iywstbdywdiy/requirements.txt
Pillow==5.1.0
tqdm==4.23.0
